Added Name#real_text_name and Name#real_search_name in order to handle the e-unlaut better. For the most part, text_name and search_name should never be seen by the user, instead use real_text_name and real_search_name.  All sorts should use sort_name.  All searches should use text_name and search_name after cleaning out the umlaut first.  This is paving the way to eventually removing all accents from search_name, and possibly allowing rather more significant differences between names that are displayed (whatever the format) and names that are used internally (with or without author).  Probably should explicitly store real_text_name and real_search_name in the database.  It's not clear to me (and never really has been) why display_name is stored in the database, and not produced on the fly.  But that's hardly a big concern in the grand scheme of things...

+++ test/unit/abstract_model_test.rb
@@ -219,15 +219,22 @@ class AbstractModelTest < UnitTestCase
       end
 
       time = obj.rss_log.modified
-      obj.destroy
-      assert_equal((num+=2), obj.rss_log.notes.split("\n").length,
-                   "#{model}.rss_log should have create, test, update, destroy, orphan lines:\n" +
-                   "<#{obj.rss_log.notes}>")
-      assert_match(/log_#{model_name}_destroyed/, obj.rss_log.notes,
-                   "#{model}.rss_log should have destroy line:\n" +
-                   "<#{obj.rss_log.notes}>")
-      assert_equal(time, obj.rss_log.modified,
-                   "#{model}.rss_log shouldn't have been touched")
+      if model == Name
+        Name.first.merge(obj)
+        assert_equal((num+=1), obj.rss_log.notes.split("\n").length,
+                     "#{model}.rss_log should have test and merge lines:\n" +
+                     "<#{obj.rss_log.notes}>")
+      else
+        obj.destroy
+        assert_equal((num+=2), obj.rss_log.notes.split("\n").length,
+                     "#{model}.rss_log should have create, test, update, destroy, orphan lines:\n" +
+                     "<#{obj.rss_log.notes}>")
+        assert_match(/log_#{model_name}_destroyed/, obj.rss_log.notes,
+                     "#{model}.rss_log should have destroy line:\n" +
+                     "<#{obj.rss_log.notes}>")
+        assert_equal(time, obj.rss_log.modified,
+                     "#{model}.rss_log shouldn't have been touched")
+      end
     end
   end
 end
